Remove deprecated class PublicationDeliveryExporter.java
