+++ src/main/java/org/rutebanken/tiamat/config/StreamingPublicationDeliveryConfig.java
@@ -15,7 +15,7 @@
 
 package org.rutebanken.tiamat.config;
 
-import org.rutebanken.tiamat.exporter.PublicationDeliveryExporter;
+import org.rutebanken.tiamat.exporter.PublicationDeliveryCreator;
 import org.rutebanken.tiamat.exporter.StreamingPublicationDelivery;
 import org.rutebanken.tiamat.exporter.TiamatFareFrameExporter;
 import org.rutebanken.tiamat.exporter.TiamatResourceFrameExporter;
@@ -49,7 +49,7 @@ public class StreamingPublicationDeliveryConfig {
     private ParkingRepository parkingRepository;
 
     @Autowired
-    private PublicationDeliveryExporter publicationDeliveryExporter;
+    private PublicationDeliveryCreator publicationDeliveryCreator;
 
     @Autowired
     private TiamatSiteFrameExporter tiamatSiteFrameExporter;
@@ -104,7 +104,7 @@ public class StreamingPublicationDeliveryConfig {
     }
 
     private StreamingPublicationDelivery createStreamingPublicationDelivery(boolean validate) throws IOException, SAXException {
-        return new StreamingPublicationDelivery(stopPlaceRepository, parkingRepository, publicationDeliveryExporter,
+        return new StreamingPublicationDelivery(stopPlaceRepository, parkingRepository, publicationDeliveryCreator,
                 tiamatSiteFrameExporter,tiamatServiceFrameExporter,tiamatFareFrameExporter,tiamatResourceFrameExporter, netexMapper, tariffZoneRepository, fareZoneRepository, topographicPlaceRepository,
                 groupOfStopPlacesRepository,groupOfTariffZonesRepository, netexIdHelper, validate, purposeOfGroupingRepository);
     }
